refactor: create infra layers to isolate the business rules of the app
- The entire layer external to the application will be passed to an infrastructure layer (infra).
- With this we will be able to isolate the application domain (business rule), from this external layer.
- This decoupling makes it possible to remove and add new technologies to the application, as needed.
- Currently, in the infra layer there are the entities, repositories, configurations and migrations of the database, all linked to TypeORM. In addition to the routes and server, coupled to the express framework.
