+++ src/modules/cars/infra/typeorm/repositories/SpecificationsRepository.ts
@@ -1,10 +1,11 @@
 import { getRepository, Repository } from 'typeorm';
 
-import { Specification } from '../../entities/Specification';
 import {
-  ISpecificationsRepository,
   ICreateSpecificationDTO,
-} from '../ISpecificationsRepository';
+  ISpecificationsRepository,
+} from '@modules/cars/repositories/ISpecificationsRepository';
+
+import { Specification } from '../entities/Specification';
 
 class SpecificationsRepository implements ISpecificationsRepository {
   private repository: Repository<Specification>;
