+++ src/modules/cars/useCases/createCategory/CreateCategoryUseCase.spec.ts
@@ -1,5 +1,5 @@
-import { AppError } from '@errors/AppError';
 import { CategoriesRepositoryInMemory } from '@modules/cars/repositories/in-memory/CategoriesRepositoryInMemory';
+import { AppError } from '@shared/errors/AppError';
 
 import { CreateCategoryUseCase } from './CreateCategoryUseCase';
