+++ src/shared/infra/http/routes/specifications.routes.ts
@@ -1,6 +1,6 @@
 import { Router } from 'express';
 
-import { CreateSpecificationController } from '../modules/cars/useCases/createSpecification/CreateSpecificationController';
+import { CreateSpecificationController } from '@modules/cars/useCases/createSpecification/CreateSpecificationController';
 
 const specificationsRoutes = Router();
